Clean up database folder
